Remove ruby version from Gemfile
Heroku defaults to ruby 2.0.0 so this line is redundant. The line caused problems on my system because I am using ruby 2.1.5 and bundler refused to run. If we need to revisit this issue at some point, this is one possible solution to allow different version of ruby in development and production https://devcenter.heroku.com/articles/ruby-versions#specifying-a-ruby-version-via-the-environment