Return graph widget to it's original state.
The graph widget was modified in 832eb39ee51c682248f6ed9ec7d304f1f03860d7 but isn't actually used for the Dark Sky widget. I'm returning it to the default state to avoid confusion about it's purpose. However, if we're not using the graph widget, maybe we should just remove it?